Extract climate data of Famine Early Warning Systems Network (FEWS NET) Land Data Assimilation System ()

A function that extract a time series of climate variables (1982-01-01T00:00:00Z–2022-04-01T00:00:00).

Usage

get_fldas(from, to, by, band, region, fun = "mean", scale = 1000)

Arguments

to, from

it's a string object,starting and final date.

by

two types of increment of the sequence by month and year.

band

name of band.

region

is a feature or feature collection.

fun

function for extract statistic zonal (count, kurtosis, max, mean, median, min, mode, percentile, std, sum, variance, first).

scale

A nominal scale in meters of the projection to work in.

Value

a tibble object with the new variables.

Details

Name of some bands.

Evap_tavg (kg m-2 s-1): Evapotranspiration.

LWdown_f_tavg (W m-2): Downward longwave radiation flux .

Lwnet_tavg (W m-2): Net longwave radiation flux.

Psurf_f_tavg (Pa): Surface pressure.

Qair_f_tavg (kg kg-1): Specific humidity.

Qg_tavg (W m-2): Soil heat flux.

Qh_tavg (W m-2): Sensible heat net flux.

Qle_tavg (W m-2): Latent heat net flux.

Qs_tavg (kg m-2 s-1): Storm surface runoff.

Qsb_tavg (kg m-2 s-1): Baseflow-groundwater runoff.

RadT_tavg (K): Surface radiative temperature.

Rainf_f_tavg (kg m-2 s-1): Total precipitation rate.

SnowCover_inst : Snow cover fraction.

SnowDepth_inst (m): Snow depth.

Snowf_tavg (kg m-2 s-1): Snowfall rate.

SoilMoi00_10cm_tavg (m^3 m-3):Soil moisture (0 - 10 cm underground).

SoilMoi10_40cm_tavg (m^3 m-3): Soil moisture (10 - 40 cm underground).

SoilMoi100_200cm_tavg (m^3 m-3): Soil moisture (100 - 200 cm underground).

SoilMoi40_100cm_tavg (m^3 m-3): Soil moisture (40 - 100 cm underground).

SoilTemp00_10cm_tavg (K): Soil temperature (0 - 10 cm underground).

SoilTemp10_40cm_tavg (K): Soil temperature (10 - 40 cm underground).

SoilTemp100_200cm_tavg (K): Soil temperature (100 - 200 cm underground).

SoilTemp40_100cm_tavg (K): Soil temperature (40 - 100 cm underground).

SWdown_f_tavg (W m-2): Surface downward shortwave radiation.

SWE_inst (kg m-2): Snow water equivalent.

Swnet_tavg (W m-2): Net shortwave radiation flux.

Tair_f_tavg (K): Near surface air temperature.

Wind_f_tavg (m s-1): Near surface wind speed.

Examples

if (FALSE) {

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region , id = 'distr' , simplify = 1000)
# 2. Extracting climate information
data <- region_ee %>% get_fldas(
  from = "2001-02-01", to = "2002-12-31",
  by = "month", band = "Qair_f_tavg", fun = "mean")
}