Extract vegetation index of MODIS

A function that extract the values of vegetation by month.

get_vegetation(from, to, band, region, fun = "count", scale = 1000)

Arguments

to, from

the starting and final range of date.

band

name of bands.

region

is a feature or feature collection.

fun

function for extract statistic zonal (count, kurtosis, max, mean, median , min, mode , percentile , std, sum, variance, first).

scale

A nominal scale in meters of the projection to work in.

Value

a sf object with the new variables.

Details

Name of some bands.

NDVI: Normalized Difference Vegetation Index.

EVI: Enhanced Vegetation Index.

SAVI: Soil Adjusted Vegetation Index.

Examples

if (FALSE) {

library(tidyverse)
library(rgee)
library(innovar)
library(sf)
ee_Initialize()

# 1. Reading a sf object
data("Peru")
region <- Peru
region_ee <- pol_as_ee(region, id = "distr", simplify = 1000)

# 2. Extracting climate information
data <- region_ee %>% get_vegetation(
  from = "2001-02-01", to = "2002-12-31", band = "NDVI", fun = "max"
)
}