module_coords

creates a data set named coords that describes the association to be considered between variables of a module to be simulated, and details to map such information into a matrix.

module_coords(
  n_var = ncol(gen_att),
  which_neg = c("1_2", "4_5", "2_3", "20_4", "30_3"),
  dis_neg_know_ass = known_neg_dis_gen,
  coo_known_neg_ass = c("4_5", "2_3", "20_4", "30_3"),
  dis_pos_know_ass = known_pos_dis_gen,
  coo_known_pos_ass = c("3_9", "4_5", "6_8", "7_8"),
  by = 1e-04
)

Arguments

n_var

number of the variables in the module to be simulated

which_neg

string with coordinates of the variables which are negativelly correlated.

dis_neg_know_ass

Distances describing negative assocaitions known a priori

coo_known_neg_ass

Coordinates that describe the positive association between two variables ex 1_2 indicates association between first and second variable

dis_pos_know_ass

Distances describing positive assocaitions known a priori

coo_known_pos_ass

Coordinates that describe the positive association between two variables ex 1_2 indicates association between first and second variable

by

parameter like in the sample function used to generate randomly unknown positive and negative associations

Value

a list with an coords a data set named coords that describes the association to be considered between variables of a module to be simulated, and details to map such information into a matrix.

Examples

module_coords(n_var=9,which_neg=c("2_1","6_2","1_9","1_6","2_3"),dis_neg_know_ass=c(-0.9300003, -0.9310004, -0.8684074),coo_known_neg_ass=c("2_1","6_2","1_9"),dis_pos_know_ass=c(0.5235988, 0.6435011, 1.1197695),coo_known_pos_ass=c("3_9","4_5","6_8"),by=0.001)
#> $coords
#>    i j pos_neg coord_ij   distance       lower     diag        upper
#> 1  1 1     pos      1_1  1.5707960 non_low_tri     diag non_uper_tri
#> 2  2 1     neg      2_1 -0.9300003     low_tri non_diag non_uper_tri
#> 3  3 1     pos      3_1  1.2630000     low_tri non_diag non_uper_tri
#> 4  4 1     pos      4_1  0.8540000     low_tri non_diag non_uper_tri
#> 5  5 1     pos      5_1  0.9180000     low_tri non_diag non_uper_tri
#> 6  6 1     neg      6_1  4.4275927     low_tri non_diag non_uper_tri
#> 7  7 1     pos      7_1  0.7360000     low_tri non_diag non_uper_tri
#> 8  8 1     pos      8_1  0.8460000     low_tri non_diag non_uper_tri
#> 9  9 1     neg      9_1 -0.8684074     low_tri non_diag non_uper_tri
#> 10 1 2     neg      1_2 -0.9300003 non_low_tri non_diag     uper_tri
#> 11 2 2     pos      2_2  1.5707960 non_low_tri     diag non_uper_tri
#> 12 3 2     neg      3_2  4.0615927     low_tri non_diag non_uper_tri
#> 13 4 2     pos      4_2  1.0200000     low_tri non_diag non_uper_tri
#> 14 5 2     pos      5_2  0.3590000     low_tri non_diag non_uper_tri
#> 15 6 2     neg      6_2 -0.9310004     low_tri non_diag non_uper_tri
#> 16 7 2     pos      7_2  0.9570000     low_tri non_diag non_uper_tri
#> 17 8 2     pos      8_2  0.4610000     low_tri non_diag non_uper_tri
#> 18 9 2     pos      9_2  0.5710000     low_tri non_diag non_uper_tri
#> 19 1 3     pos      1_3  1.2630000 non_low_tri non_diag     uper_tri
#> 20 2 3     neg      2_3  4.0615927 non_low_tri non_diag     uper_tri
#> 21 3 3     pos      3_3  1.5707960 non_low_tri     diag non_uper_tri
#> 22 4 3     pos      4_3  0.9310000     low_tri non_diag non_uper_tri
#> 23 5 3     pos      5_3  0.9040000     low_tri non_diag non_uper_tri
#> 24 6 3     pos      6_3  0.9510000     low_tri non_diag non_uper_tri
#> 25 7 3     pos      7_3  0.1720000     low_tri non_diag non_uper_tri
#> 26 8 3     pos      8_3  1.1780000     low_tri non_diag non_uper_tri
#> 27 9 3     pos      9_3  0.5235988     low_tri non_diag non_uper_tri
#> 28 1 4     pos      1_4  0.8540000 non_low_tri non_diag     uper_tri
#> 29 2 4     pos      2_4  1.0200000 non_low_tri non_diag     uper_tri
#> 30 3 4     pos      3_4  0.9310000 non_low_tri non_diag     uper_tri
#> 31 4 4     pos      4_4  1.5707960 non_low_tri     diag non_uper_tri
#> 32 5 4     pos      5_4  0.6435011     low_tri non_diag non_uper_tri
#> 33 6 4     pos      6_4  1.5610000     low_tri non_diag non_uper_tri
#> 34 7 4     pos      7_4  1.0060000     low_tri non_diag non_uper_tri
#> 35 8 4     pos      8_4  0.1790000     low_tri non_diag non_uper_tri
#> 36 9 4     pos      9_4  1.1400000     low_tri non_diag non_uper_tri
#> 37 1 5     pos      1_5  0.9180000 non_low_tri non_diag     uper_tri
#> 38 2 5     pos      2_5  0.3590000 non_low_tri non_diag     uper_tri
#> 39 3 5     pos      3_5  0.9040000 non_low_tri non_diag     uper_tri
#> 40 4 5     pos      4_5  0.6435011 non_low_tri non_diag     uper_tri
#> 41 5 5     pos      5_5  1.5707960 non_low_tri     diag non_uper_tri
#> 42 6 5     pos      6_5  0.6130000     low_tri non_diag non_uper_tri
#> 43 7 5     pos      7_5  0.8860000     low_tri non_diag non_uper_tri
#> 44 8 5     pos      8_5  1.5170000     low_tri non_diag non_uper_tri
#> 45 9 5     pos      9_5  0.4400000     low_tri non_diag non_uper_tri
#> 46 1 6     neg      1_6  4.4275927 non_low_tri non_diag     uper_tri
#> 47 2 6     neg      2_6 -0.9310004 non_low_tri non_diag     uper_tri
#> 48 3 6     pos      3_6  0.9510000 non_low_tri non_diag     uper_tri
#> 49 4 6     pos      4_6  1.5610000 non_low_tri non_diag     uper_tri
#> 50 5 6     pos      5_6  0.6130000 non_low_tri non_diag     uper_tri
#> 51 6 6     pos      6_6  1.5707960 non_low_tri     diag non_uper_tri
#> 52 7 6     pos      7_6  0.3020000     low_tri non_diag non_uper_tri
#> 53 8 6     pos      8_6  1.1197695     low_tri non_diag non_uper_tri
#> 54 9 6     pos      9_6  1.3690000     low_tri non_diag non_uper_tri
#> 55 1 7     pos      1_7  0.7360000 non_low_tri non_diag     uper_tri
#> 56 2 7     pos      2_7  0.9570000 non_low_tri non_diag     uper_tri
#> 57 3 7     pos      3_7  0.1720000 non_low_tri non_diag     uper_tri
#> 58 4 7     pos      4_7  1.0060000 non_low_tri non_diag     uper_tri
#> 59 5 7     pos      5_7  0.8860000 non_low_tri non_diag     uper_tri
#> 60 6 7     pos      6_7  0.3020000 non_low_tri non_diag     uper_tri
#> 61 7 7     pos      7_7  1.5707960 non_low_tri     diag non_uper_tri
#> 62 8 7     pos      8_7  0.9600000     low_tri non_diag non_uper_tri
#> 63 9 7     pos      9_7  0.6690000     low_tri non_diag non_uper_tri
#> 64 1 8     pos      1_8  0.8460000 non_low_tri non_diag     uper_tri
#> 65 2 8     pos      2_8  0.4610000 non_low_tri non_diag     uper_tri
#> 66 3 8     pos      3_8  1.1780000 non_low_tri non_diag     uper_tri
#> 67 4 8     pos      4_8  0.1790000 non_low_tri non_diag     uper_tri
#> 68 5 8     pos      5_8  1.5170000 non_low_tri non_diag     uper_tri
#> 69 6 8     pos      6_8  1.1197695 non_low_tri non_diag     uper_tri
#> 70 7 8     pos      7_8  0.9600000 non_low_tri non_diag     uper_tri
#> 71 8 8     pos      8_8  1.5707960 non_low_tri     diag non_uper_tri
#> 72 9 8     pos      9_8  0.7170000     low_tri non_diag non_uper_tri
#> 73 1 9     neg      1_9 -0.8684074 non_low_tri non_diag     uper_tri
#> 74 2 9     pos      2_9  0.5710000 non_low_tri non_diag     uper_tri
#> 75 3 9     pos      3_9  0.5235988 non_low_tri non_diag     uper_tri
#> 76 4 9     pos      4_9  1.1400000 non_low_tri non_diag     uper_tri
#> 77 5 9     pos      5_9  0.4400000 non_low_tri non_diag     uper_tri
#> 78 6 9     pos      6_9  1.3690000 non_low_tri non_diag     uper_tri
#> 79 7 9     pos      7_9  0.6690000 non_low_tri non_diag     uper_tri
#> 80 8 9     pos      8_9  0.7170000 non_low_tri non_diag     uper_tri
#> 81 9 9     pos      9_9  1.5707960 non_low_tri     diag non_uper_tri
#>    known_priori        sin
#> 1      not_know  1.0000000
#> 2         known -0.8016201
#> 3      not_know  0.9530035
#> 4      not_know  0.7539143
#> 5      not_know  0.7943884
#> 6      not_know -0.9597189
#> 7      not_know  0.6713287
#> 8      not_know  0.7486345
#> 9         known -0.7633010
#> 10        known -0.8016201
#> 11     not_know  1.0000000
#> 12     not_know -0.7956016
#> 13     not_know  0.8521080
#> 14     not_know  0.3513382
#> 15        known -0.8022176
#> 16     not_know  0.8174673
#> 17     not_know  0.4448439
#> 18     not_know  0.5404737
#> 19     not_know  0.9530035
#> 20     not_know -0.7956016
#> 21     not_know  1.0000000
#> 22     not_know  0.8022174
#> 23     not_know  0.7858071
#> 24     not_know  0.8139968
#> 25     not_know  0.1711532
#> 26     not_know  0.9238423
#> 27        known  0.5000000
#> 28     not_know  0.7539143
#> 29     not_know  0.8521080
#> 30     not_know  0.8022174
#> 31     not_know  1.0000000
#> 32        known  0.6000000
#> 33     not_know  0.9999520
#> 34     not_know  0.8446976
#> 35     not_know  0.1780456
#> 36     not_know  0.9086335
#> 37     not_know  0.7943884
#> 38     not_know  0.3513382
#> 39     not_know  0.7858071
#> 40        known  0.6000000
#> 41     not_know  1.0000000
#> 42     not_know  0.5753238
#> 43     not_know  0.7745479
#> 44     not_know  0.9985533
#> 45     not_know  0.4259395
#> 46     not_know -0.9597189
#> 47        known -0.8022176
#> 48     not_know  0.8139968
#> 49     not_know  0.9999520
#> 50     not_know  0.5753238
#> 51     not_know  1.0000000
#> 52     not_know  0.2974303
#> 53        known  0.9000000
#> 54     not_know  0.9797081
#> 55     not_know  0.6713287
#> 56     not_know  0.8174673
#> 57     not_know  0.1711532
#> 58     not_know  0.8446976
#> 59     not_know  0.7745479
#> 60     not_know  0.2974303
#> 61     not_know  1.0000000
#> 62     not_know  0.8191916
#> 63     not_know  0.6202019
#> 64     not_know  0.7486345
#> 65     not_know  0.4448439
#> 66     not_know  0.9238423
#> 67     not_know  0.1780456
#> 68     not_know  0.9985533
#> 69        known  0.9000000
#> 70     not_know  0.8191916
#> 71     not_know  1.0000000
#> 72     not_know  0.6571263
#> 73        known -0.7633010
#> 74     not_know  0.5404737
#> 75        known  0.5000000
#> 76     not_know  0.9086335
#> 77     not_know  0.4259395
#> 78     not_know  0.9797081
#> 79     not_know  0.6202019
#> 80     not_know  0.6571263
#> 81     not_know  1.0000000
#> 
#> $which_negative
#>  [1] "2_1" "6_2" "1_9" "1_6" "2_3" "1_2" "2_6" "9_1" "6_1" "3_2"
#>