set_sim_copula

set_sim_copula

set_sim_copula(
  d = 9,
  lower_tri_corr = soil_corr[lower.tri(soil_corr, diag = FALSE)],
  n_cont_var = ncol(soil_num),
  cont_var_par = list(list(lambda = 40), list(df = 30, ncp = 3), list(shape = 40, scale =
    10), list(rate = 30)),
  n_unique = length(unique(na.omit(soil_att_ex$siteid))),
  mar_cont_dists = c("pois", "chisq", "gamma", "exp"),
  var_names = colnames(soil_att_ex)
)

Arguments

d

dimension of copula to be set. The same as the number of variables inside a module

lower_tri_corr

lower triangular part of a valida correlation matrix

n_cont_var

number of continuous variables inside the module and which will be described by specific marginal distributions

cont_var_par

a list of lists containing names and values of the parameters of the

n_unique

number of sites for which the simulations should be made

mar_cont_dists

a vector of strings containing strings with the names of the marginal distributions

var_names

name of the variables inside the moduls

Value

a list containing the copula settled for the simulations its mvdc component and the simulated continuos values. some of them will be siliced to generate categorical components